Add dispatch
Introduce a message handler interface that allows closures to be
registered with an object that handles dispatch. The message handler can
be invoked in an application's main loop to consume messages and route
them to the registered callbacks. Callbacks are routed based on their
object, interface, and member paths used when registered.

Also some basic validation is performed to guarantee that all required
parts for the message for a given message type are valid and present.
Error messages to deliver to the bus are generated when validation
fails.

This currently only handles signals and method calls. It should be extended
to support method returns and errors to be dispatched by reply_serial.
There is also not yet a deregistration interface.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>