dispatch: add Method and Signal handler wrapper macros
These macros bind Rust functions to generic closures which can be used
with the message dispatcher.

For both signals and methods, the glue code validates the arguments in
the message body received from the bus and will return error responses
to the bus without invoking the callback.

Methods may optionally return a tuple of Encodables which will be
inserted in the body of the autogenerated method call response message.
The choice of a method to produce an autogenerated response is decided
at run time via an Option return. This allows method calls to
dynamically make the decision to defer the response message.

Method return arguments are guaranteed to be Encodable at compile time.
Since the methods have a signature known at compile time, this also
ensures a single method will always have the same return signature.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>